requirements.txt
# Open-Monitor v3.0 - Dependencies
# Flask & Extensions
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1

# Security
bcrypt==4.1.1
cryptography==41.0.7
PyJWT==2.8.0
python-dotenv==1.0.0
email-validator==2.1.0

# HTTP & Async
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# Scheduling
APScheduler==3.10.4
celery==5.3.6

# Data Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
WTForms==3.1.1

# PDF Generation
WeasyPrint==60.1

# OpenAI Integration
openai==1.6.1

# Web Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0
